gdbrel/__init__.py: Make sure the remote's URL is correct
... or, said differently, that the remote's URL matches the one we have
in the config file. In particular, when re-using an existing clone of
the repository, make sure that the URL of that existing clone matches
the URL defined in our config file.

This is important, especially when that URL is temporarily changed
as a way to perform local testing. When doing such testing, one does
not want to use that clone and push with the wrong URL!